[FP-59] Refactor: 체험 상세 페이지 테블릿 캘린더 예약하기 post 로직 위치 변경
