+++ src/components/common/FloatingBox/TabletComponents/main.tsx
@@ -1,6 +1,9 @@
 import Button from '@/components/common/Button';
 import Counter from '@/components/common/Counter';
 import Modal from '@/components/common/Modal';
+import { useActivityReservationMutation } from '@/hooks/useReservationMutation';
+import { activityIdAtom, formSubmitDataAtom } from '@/store/activityDetailsAtom';
+import { useAtomValue } from 'jotai';
 import { useState } from 'react';
 
 const Main = () => {
@@ -9,6 +12,14 @@ const Main = () => {
   const handleDateSelectModal = () => {
     // TODO: 모달 여닫는 기능 구현
     setIsOpenModal((prev) => !prev);
+  };
+
+  const { submitReservation } = useActivityReservationMutation();
+  const activityId = useAtomValue(activityIdAtom);
+  const submitFormData = useAtomValue(formSubmitDataAtom);
+
+  const handleReservationFormSubmit = () => {
+    submitReservation(activityId, submitFormData);
   };
 
   return (
@@ -23,7 +34,9 @@ const Main = () => {
         <div className="mb-32">
           <Counter />
         </div>
-        <Button.Default className="h-56 w-203">예약하기</Button.Default>
+        <Button.Default className="h-56 w-203" onClick={handleReservationFormSubmit}>
+          예약하기
+        </Button.Default>
       </main>
     </div>
   );
